Move plan_mapped_price enum
