+++ real-main/app/models/appstore/enums.py
@@ -1,3 +1,7 @@
+from decimal import Decimal
+from enum import Enum
+
+
 class AppStoreSubscriptionStatus:
     # Note: we do not have a grace period configured at this time
     ACTIVE = 'ACTIVE'
@@ -11,3 +15,7 @@ class PricePlan:
     SUBSCRIPTION_DIAMOND = 'SUBSCRIPTION_DIAMOND'
 
     _ALL = SUBSCRIPTION_DIAMOND
+
+
+class PlanMappedPrice(Enum):
+    SUBSCRIPTION_DIAMOND = Decimal('0.99')
