--- real-main/app/models/appstore/plan_mapped_price.py
@@ -1,6 +1,0 @@
-from decimal import Decimal
-from enum import Enum
-
-
-class PlanMappedPrice(Enum):
-    SUBSCRIPTION_DIAMOND = Decimal('0.99')
