+++ real-main/app/models/user/manager.py
@@ -11,8 +11,7 @@ from app import models
 from app.clients import AmplitudeClient
 from app.mixins.base import ManagerBase
 from app.mixins.trending.manager import TrendingManagerMixin
-from app.models.appstore.enums import AppStoreSubscriptionStatus
-from app.models.appstore.plan_mapped_price import PlanMappedPrice
+from app.models.appstore.enums import AppStoreSubscriptionStatus, PlanMappedPrice
 from app.models.card.templates import ContactJoinedCardTemplate, UserNewDatingMatchesTemplate
 from app.models.follower.enums import FollowStatus
 from app.models.post.enums import PostStatus
